[-] BciNumber enumeration was removed
[*] All entity BciNumber replaced by quint16 and xml config parser map
[*] Unpacking program data replaced to coma init step

+++ src/s2/configv.cpp
@@ -4,7 +4,7 @@
 
 #include <QDebug>
 
-DataTypes::DataRecV ConfigV::getRecord(unsigned int id)
+DataTypes::DataRecV ConfigV::getRecord(quint16 id)
 {
     auto result = std::find_if(
         std::cbegin(config), std::cend(config), [id](const auto &record) { return (id == record.getId()); });
