+++ src/s2/configv.h
@@ -7,7 +7,7 @@ class ConfigV
 public:
     ConfigV() = default;
 
-    DataTypes::DataRecV getRecord(unsigned int id);
+    DataTypes::DataRecV getRecord(quint16 id);
     void setRecordValue(const DataTypes::DataRecV &record);
 
     const QList<DataTypes::DataRecV> &getConfig() const;
